requirements.txt
aiobotocore==2.4.2
aiofiles==22.1.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.8.1
amqp==5.1.1
annotated-types==0.5.0
ansible-base==2.10.8
anyio==3.6.2
apache-airflow==2.5.0
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-postgres==2.4.0
apache-airflow-providers-sqlite==3.3.1
apache-libcloud==3.2.0
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.27.1
azure-storage-blob==12.16.0
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==4.1.0
blinker==1.5
boto3==1.26.70
botocore==1.27.59
Brlapi==0.8.3
bs4==0.0.1
cached-property==1.5.2
cachelib==0.9.0
cachetools==5.2.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.9.24
certipy==0.1.3
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
cloudpickle==2.2.1
cmake==3.26.4
cmarkgfm==2022.10.27
colorama==0.4.6
colorlog==4.8.0
comm==0.1.2
command-not-found==0.3
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
contourpy==1.0.7
cron-descriptor==1.2.32
croniter==1.3.8
crypto==1.4.1
cryptography==38.0.4
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
databricks-cli==0.17.7
datasketch==1.5.8
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.1.1
distlib==0.3.6
distro==1.8.0
distro-info==1.0
dnspython==2.2.1
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
dotmap==1.3.30
duplicity==0.8.21
EbookLib==0.17.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
exrex==0.10.5
fake-useragent==0.1.14
Faker==15.3.1
fastDamerauLevenshtein==1.0.7
fasteners==0.18
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flower==1.2.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.10.2
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.21.0
httpx==0.24.1
huggingface-hub==0.15.1
humanize==4.4.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
install==1.3.5
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-telemetry==0.1.0
jupyter-ydoc==0.2.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_fileid==0.6.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterhub==4.0.0
jupyterlab==3.6.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
keycloak==3.1.3
keyring==23.11.0
kiwisolver==1.4.4
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.18
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lazy-object-proxy==1.8.0
linkify-it-py==2.0.0
lit==16.0.5.post0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
migra==3.0.1663481299
mistune==2.0.4
mlflow==2.3.1
monotonic==1.6
more-itertools==9.0.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
Naked==0.1.32
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
neovim==0.3.1
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
ntlm-auth==1.5.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthenticator==15.1.0
oauthlib==3.2.2
olefile==0.46
packaging==21.3
pamela==1.0.0
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.12.0
parse==1.19.0
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.8.3
platformdirs==2.5.3
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==4.21.9
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==2.1.1
protonvpn-nm-lib==3.14.0
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybloomfilter==1.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.15.0
pycups==2.0.1
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
PyDrive==1.3.1
pyee==9.0.4
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.6.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyRFC3339==1.1
pyrsistent==0.19.2
pyspark==3.3.1
pysqlite3-binary==0.5.0
python-apt==2.3.0+ubuntu2.1
python-daemon==2.3.2
python-dateutil==2.8.2
python-debian==0.1.48
python-dotenv==0.21.0
python-gnupg==0.5.0
python-jose==3.3.0
python-json-logger==2.0.4
python-keycloak==2.16.0
python-nvd3==0.15.0
python-slugify==7.0.0
python-telegram-bot==20.4
pythondialog==3.5.3
pytz==2022.6
pytzdata==2020.1
pywinrm==0.4.3
pyxdg==0.28
PyYAML==5.4.1
pyzmq==25.0.0
querystring-parser==1.2.4
rdt==0.2.1
readme-renderer==37.3
redis==4.5.4
regex==2022.10.31
requests==2.28.1
requests-html==0.10.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
roman==3.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2023.3.0
s3transfer==0.6.0
safetensors==0.3.1
schemainspect==3.1.1663587362
scikit-learn==1.2.2
scipy==1.9.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
selinux==3.3
Send2Trash==1.8.0
setproctitle==1.3.2
shellescape==3.8.1
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlbag==0.1.1617247075
sqlparse==0.4.3
ssh-import-id==5.11
stack-data==0.6.2
swagger-ui-bundle==0.0.9
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.1.1
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.13.3
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
triton==2.0.0
twine==4.0.1
typing_extensions==4.7.1
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.1
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.16.6
w3lib==2.0.1
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
xdg==5.1.1
xkit==0.0.0
XlsxWriter==3.0.3
xmltodict==0.13.0
y-py==0.5.5
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.10.0
